+++ hw2/ExplorerRay/q1/line.hpp
@@ -11,10 +11,10 @@ public:
     Line &operator=(Line       &&);
     Line(size_t size);
     size_t size() const;
-    float const &x(size_t it) const;
-    float &x(size_t it);
-    float const &y(size_t it) const;
-    float &y(size_t it);
+    double const &x(size_t it) const;
+    double &x(size_t it);
+    double const &y(size_t it) const;
+    double &y(size_t it);
 private:
-    std::vector<std::pair<float, float>> m_points; // 2D points
+    std::vector<std::pair<double, double>> m_points; // 2D points
 }; /* end class Line */
